IGNITE-19501 / introduce TableEventParameters
